+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 - **Forum Changes**
   - Forum topics now track views and have better URL structure
   - Users can upvote, downvote, and reply directly to other forum posts
+  - Users now receive in-app notifications when a new reply is added to a forum
+    topic to which they are subscribed
 - **Search Added**
   - A site-wide search has been added that searches through blog posts and forum
     topics